simple-backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
gunicorn==21.2.0
aiohttp==3.9.0
asyncio
numpy==1.24.3
hvac==1.2.1
reportlab==4.0.4
psycopg2-binary==2.9.7
asyncpg==0.29.0
bcrypt==4.0.1
PyJWT==2.8.0

# Async job queue
rq==1.15.1
redis==5.0.1

# Observability dependencies
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-aiohttp==0.41b0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-prometheus==0.41b0
opentelemetry-propagator-b3==1.20.0
prometheus-client==0.18.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
pytest-mock==3.11.1
coverage==7.3.0